Move HelperMethods.scala to javalib, make package private
